refactor: 파일 분리 및 구조 개선
