+++ app/src/main/java/com/example/letssopt/presentation/LoginActivity.kt
@@ -1,4 +1,4 @@
-package com.example.letssopt
+package com.example.letssopt.presentation
 
 import android.content.Intent
 import android.os.Bundle
@@ -37,12 +37,13 @@ import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.letssopt.ui.theme.Background
-import com.example.letssopt.ui.theme.LETSSOPTTheme
-import com.example.letssopt.ui.theme.PrimaryRed
-import com.example.letssopt.ui.theme.Surface
-import com.example.letssopt.ui.theme.TextPrimary
-import com.example.letssopt.ui.theme.TextSecondary
+import com.example.letssopt.R
+import com.example.letssopt.designsystem.theme.Background
+import com.example.letssopt.designsystem.theme.LETSSOPTTheme
+import com.example.letssopt.designsystem.theme.PrimaryRed
+import com.example.letssopt.designsystem.theme.Surface
+import com.example.letssopt.designsystem.theme.TextPrimary
+import com.example.letssopt.designsystem.theme.TextSecondary
 
 class LoginActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
